+++ src/server/scripts/Commands/cs_modify.cpp
@@ -1020,8 +1020,8 @@ public:
             }
             else
             {
-                uint32 moneyToAddMsg = moneyToAdd * -1;
-                if (newmoney > int64(MAX_MONEY_AMOUNT))
+                uint64 moneyToAddMsg = moneyToAdd * -1;
+                if (newmoney > static_cast<int64>(MAX_MONEY_AMOUNT))
                     newmoney = MAX_MONEY_AMOUNT;
 
                 handler->PSendSysMessage(LANG_YOU_TAKE_MONEY, moneyToAddMsg, handler->GetNameLink(target).c_str());
